Display relative time.
Frontend has datetime as YYYY-MM-DDThh:mm:ss string format, so it
needs to be translated to human-friendly format. Another time
format would be good, for example, 2018-09-20 12:34:56, however
the most practices users want to see relative time. The heartbeats
for devices and jobs are good examples. Show details button is
going to be introduced soon and in that UI users still can see
the raw time value.

Test: go/vtslab-schedule-dev
Bug: 74575555
Change-Id: I996ac5ea23912a8e5ee3b842ec147fa3df319a47